Update mounting system
Adds keys "AdditionalContentRoot" and "AdditionalContentPaths".
AdditionalContentRoot is the root mod folder from which we are mounting, this gets mounted directly for loose file search.
AdditionaContentPaths accepts the macro |content_root| to prefix AdditionalContentRoot to the path that was provided, and is parsed identically to SearchPaths.

AdditionalContentRoot is required if AdditionalContentId is provided.